move property related functions to reactjs
