+++ packages/uxpin-merge-cli/src/framework/reactjs/experimentation/server/createNewPageContent.ts
@@ -17,7 +17,9 @@ const ELEMENT_PLACEMENT_DISTANCE:number = 500;
 const ELEMENT_Y:number = 50;
 
 export function createNewPageContent({ revisionId }:PageContentContext, metadata:DesignSystemSnapshot):PageContent {
-  const introPageContent:any = require('./introPageContent.json');
+  const introPageContent:any = require(
+    '../../../../steps/experimentation/server/common/page/content/introPageContent.json',
+  );
   const introComponents:ComponentDefinition[] = findIntroComponents(metadata);
   const designSystemId:string = getDesignSystemId(revisionId);
