+++ packages/uxpin-merge-cli/src/framework/reactjs/index.ts
@@ -7,8 +7,10 @@ import { decorateWithPresets } from './serialization/presets/decorateWithPresets
 import { isDefaultExportedForwardRef } from './serialization/typescript/component/findExportedFunctionWithReactForwardRef';
 import { getComponentDeclaration } from './serialization/typescript/component/getComponentDeclaration';
 import { getSerializationContext } from './serialization/typescript/context/getSerializationContext';
+import { KNOWN_TYPES_MAP } from './serialization/typescript/property/type/node/serializeKnownPropertyType';
 
 export const reactModule:FrameworkModule = {
+  KNOWN_PROPERTY_TYPES_MAP: KNOWN_TYPES_MAP,
   PresetFileGenerator,
   createNewPageContent,
   decorateWithPresets,
