+++ packages/uxpin-merge-cli/src/framework/reactjs/serialization/typescript/property/symbol/convertMethodSignatureSymbolToPropertyDefinition.ts
@@ -1,5 +1,5 @@
-import { ParsedComponentProperty } from '../../../ComponentPropertyDefinition';
-import { TSSerializationContext } from '../../context/getSerializationContext';
+import { ParsedComponentProperty } from '../../../../../../steps/serialization/component/implementation/ComponentPropertyDefinition';
+import { TSSerializationContext } from '../../../../../../steps/serialization/component/implementation/typescript/context/getSerializationContext';
 import { getJSDocDocumentation } from './getJSDocDocumentation';
 import { getPropertyCustomDescriptors } from './getPropertyCustomDescriptiors';
 import { getPropertyName } from './getPropertyName';
