+++ packages/uxpin-merge-cli/src/framework/reactjs/serialization/typescript/property/symbol/getJSDocDocumentation.ts
@@ -1,5 +1,5 @@
 import * as ts from 'typescript';
-import { TSSerializationContext } from '../../context/getSerializationContext';
+import { TSSerializationContext } from '../../../../../../steps/serialization/component/implementation/typescript/context/getSerializationContext';
 import { ReactPropertySymbol } from './ReactPropertySymbol';
 
 export function getJSDocDocumentation(context:TSSerializationContext, propSymbol:ReactPropertySymbol):string {
