+++ packages/uxpin-merge-cli/src/framework/reactjs/serialization/typescript/property/symbol/getPropertyCustomDescriptiors.ts
@@ -1,6 +1,6 @@
-import { isValidDescriptor } from '../../../../props/parsing/descriptors/isValidDescriptor';
-import { ParsedPropertyDescriptors } from '../../../ComponentPropertyDefinition';
-import { parseTags } from '../../../javascript/props/parseTags';
+import { ParsedPropertyDescriptors } from '../../../../../../steps/serialization/component/implementation/ComponentPropertyDefinition';
+import { parseTags } from '../../../../../../steps/serialization/component/implementation/javascript/props/parseTags';
+import { isValidDescriptor } from '../../../../../../steps/serialization/component/props/parsing/descriptors/isValidDescriptor';
 import { ReactPropertySymbol } from './ReactPropertySymbol';
 
 export function getPropertyCustomDescriptors(
