+++ packages/uxpin-merge-cli/src/framework/reactjs/serialization/typescript/property/type/getPropertiesFromType.ts
