+++ packages/uxpin-merge-cli/src/framework/reactjs/serialization/typescript/property/type/node/convertTypeToPropertyType.ts
@@ -1,13 +1,13 @@
 import * as ts from 'typescript';
-import { PropertyType } from '../../../../ComponentPropertyDefinition';
-import { TSSerializationContext } from '../../../context/getSerializationContext';
-import { isBooleanLike } from '../checker/isBooleanLike';
-import { isCallable } from '../checker/isCallable';
-import { isEnum } from '../checker/isEnum';
-import { isKnownPropertyType } from '../checker/isKnownPropertyType';
-import { isLiteral } from '../checker/isLiteral';
-import { isObjectLike } from '../checker/isObjectLike';
-import { isUnion } from '../checker/isUnion';
+import { PropertyType } from '../../../../../../../steps/serialization/component/implementation/ComponentPropertyDefinition';
+import { TSSerializationContext } from '../../../../../../../steps/serialization/component/implementation/typescript/context/getSerializationContext';
+import { isBooleanLike } from '../../../../../../../steps/serialization/component/implementation/typescript/property/type/checker/isBooleanLike';
+import { isCallable } from '../../../../../../../steps/serialization/component/implementation/typescript/property/type/checker/isCallable';
+import { isEnum } from '../../../../../../../steps/serialization/component/implementation/typescript/property/type/checker/isEnum';
+import { isKnownPropertyType } from '../../../../../../../steps/serialization/component/implementation/typescript/property/type/checker/isKnownPropertyType';
+import { isLiteral } from '../../../../../../../steps/serialization/component/implementation/typescript/property/type/checker/isLiteral';
+import { isObjectLike } from '../../../../../../../steps/serialization/component/implementation/typescript/property/type/checker/isObjectLike';
+import { isUnion } from '../../../../../../../steps/serialization/component/implementation/typescript/property/type/checker/isUnion';
 import { serializeAsUnsupportedType } from './serializeAsUnsupportedType';
 import { serializeEnumType } from './serializeEnumType';
 import { serializeKnownPropertyType } from './serializeKnownPropertyType';
