+++ packages/uxpin-merge-cli/src/framework/reactjs/serialization/typescript/property/type/node/serializeUnionType.ts
@@ -1,6 +1,6 @@
 import * as ts from 'typescript';
-import { PropertyType } from '../../../../ComponentPropertyDefinition';
-import { TSSerializationContext } from '../../../context/getSerializationContext';
+import { PropertyType } from '../../../../../../../steps/serialization/component/implementation/ComponentPropertyDefinition';
+import { TSSerializationContext } from '../../../../../../../steps/serialization/component/implementation/typescript/context/getSerializationContext';
 import { convertTypeToPropertyType } from './convertTypeToPropertyType';
 
 export function serializeUnionType(context:TSSerializationContext, type:ts.UnionType):PropertyType<'union'> {
